evaluate in float32

+++ examples/sg_to_image/train_sg_to_image_lora.py
@@ -1200,7 +1200,7 @@ def main():
             args.pretrained_model_name_or_path,
             unet=accelerator.unwrap_model(unet),
             revision=args.revision,
-            torch_dtype=weight_dtype,
+            torch_dtype=torch.float32,
         )
         pipeline = pipeline.to(accelerator.device)
         pipeline.set_progress_bar_config(disable=True)
